==> 03_pca.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 12, 6
from sklearn.decomposition import PCA
from sklearn.preprocessing import scale

d = pd.read_csv('tmp/features_02.csv')

e = DataFrame({'VisitNumber': d.VisitNumber})
for c in d.columns:
    if c.startswith('DepartmentDescription_'):
        e[c] = d[c]
    if c.startswith('Weekday_'):
        e[c] = d[c].astype(float)
    if c.startswith('ScanCount'):
        e[c] = d[c].astype(float)
del e['VisitNumber']

N_COMPONENTS = [10, 20, 30, 50]

for n in N_COMPONENTS:
    f = e.copy()
    pca = PCA(n_components=n)
    f = pca.fit_transform(f)
    #print pca.explained_variance_ratio_
    g = scale(f)
    h = DataFrame(data=g, columns=[('pc_'+str(i)) for i in range(1, n+1)])
    h['VisitNumber'] = d.VisitNumber
    for c in d.columns:
        if c.startswith('Trip'):
            h[c] = d[c]
    h.to_csv('tmp/pca{0}_02.csv'.format(n), index=False)
