==> 08_xgb_multi_exp.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame, Series
import matplotlib.pyplot as plt
from pylab import rcParams
%pylab inline
rcParams['figure.figsize'] = 12, 6
import xgboost as xgb
from xgboost import XGBClassifier
from sklearn.metrics import roc_auc_score
from sklearn.grid_search import GridSearchCV

d = pd.read_csv('tmp/features_02.csv')
d = d[d.TripType_0 == 0]
del d['TripType_0']

tts = d.TripType.unique().tolist()
orig_to_new = {}
new_to_orig = {}
for i in range(len(tts)):
    orig_to_new[tts[i]] = i
    new_to_orig[i] = tts[i]

del d['Weekday']
del d['Weekday_nan']
lim = int(0.7 * len(d))
dtr = d[:lim]
dho = d[lim:]
tr = DataFrame({'VisitNumber': dtr.VisitNumber})
trl = DataFrame({'VisitNumber': dtr.VisitNumber})
ho = DataFrame({'VisitNumber': dho.VisitNumber})
hol = DataFrame({'VisitNumber': dho.VisitNumber})
for c in d.columns:
    if (c.startswith('ScanCount') or
        c.startswith('DepartmentDescription') or
        c.startswith('Weekday')):
        tr[c] = dtr[c].astype(float)
        ho[c] = dho[c].astype(float)
trl['TripType'] = dtr.TripType.apply(lambda t: orig_to_new[t]).astype(float)
hol['TripType'] = dho.TripType.apply(lambda t: orig_to_new[t]).astype(float)

xg_train = xgb.DMatrix(tr, label=trl.TripType)
xg_test = xgb.DMatrix(ho, label=hol.TripType)

param = {}
param['objective'] = 'multi:softmax'
param['max_depth'] = 25
param['silent'] = 1
param['n_estimators'] = 2000
param['learning_rate'] = 0.05
param['nthread'] = 4
param['num_class'] = len(tts)

watchlist = [(xg_train,'train'), (xg_test, 'test')]
num_round = 500
bst = xgb.train(param, xg_train, num_round, watchlist)
pred = bst.predict( xg_test )
print ('predicting, classification error=%f'
       % (sum( int(pred[i]) != test_Y[i] for i in range(len(test_Y))) / float(len(test_Y)) ))

p = DataFrame({'TripType': pred})
p.TripType = p.TripType.apply(lambda t: new_to_orig[t])
p = pd.get_dummies(p, columns=['TripType'])
p.index = pd.Index(np.arange(len(p)))
# print len(p)
# print len(p.columns)
# print p.columns
y_true = pd.get_dummies(hol.astype(int), columns=['TripType'])
y_true.index = pd.Index(np.arange(len(y_true)))
del y_true['VisitNumber']
# print len(y_true)
# print len(y_true.columns)
# print y_true.columns
p2 = DataFrame({}, index=y_true.index)
for c in y_true.columns:
    if c in p.columns:
        p2[c] = p[c]
        #print c, 'foo'
    else:
        p2[c] = np.zeros_like(y_true[c])
        #print c, 'bar'
z = y_true * p2
print z.sum().sum()
#z = z.apply(np.log)
#print z.sum().sum()
# z = z.replace([-inf], 0)
# print -z.sum().sum()/len(y_true)
